Add Vial support to ProjectD Tinker 65 ANSI (#895)
* Add support for ProjectD Tinker65 ANSI

Reduced nr. of layers from 13 to 4 and disabled some features to save space, because otherwise wouldn't fit EPROM

* All features enabled by default

All features now enabled by default, and only nr. of layers reduced to make fit. Can uncomment lines in rules.mk to squeeze a bit more free space out if features are not needed.

* Fix nr. of layers in keyboard.json

When uploading to github for the pull request, forgot to edit the keyboard.json to decrease the number of layers. Only did that in my local offline clone.

* Moved layer reduction to vial/config.h and reverted main keyboard.json

Reverted main keyboard.json back to default for upstream QMK compatibility. Put redefinition of layers back in vial/config.h but added a #undef to get rid of the definition from keyboard.json that caused this method to not build.